refactor(granola): extract shared granola_core; both copies import it (MYC-1529)
Collapse the Granola Public-API contract into ONE module so the substrate
exporter and the private vault's granola_export.py cannot drift. The contract
(auth, api_get, list_notes, format_transcript, safe_filename, write_transcript_md,
state, incremental window) now lives in scripts/granola_core.py; granola_sync.py
re-exports it as a thin CLI.

Prevents bug class DISCIPLINE-FIX-NOT-PORTED-TO-SUBSTRATE: the personal copy was
migrated to the Public API three weeks before the substrate copy, which silently
exported nothing in between. The two copies had already micro-drifted
(_speaker_label, frontmatter, User-Agent).

tests/integration/test_granola_core_shared.sh asserts granola_sync re-exports the
SAME objects from granola_core (identity == the anti-drift guarantee + its own
negative control), plus the extra_frontmatter variation point and the window
math. Wired into scripts/ci.sh. Existing test_granola_sync_offline.sh stays green.
