+++ scripts/ci.sh
@@ -134,6 +134,7 @@ INTEGRATION_TESTS=(
   test_connector_liveness_watchdog
   test_connector_liveness_surface
   test_granola_sync_offline
+  test_granola_core_shared
   test_agent_memory_link
   test_open_core_boundary
   test_audited_content_injection_scan
